Experimental multiple prims on a single attachment point support.
It's difficult to extend the existing "_ap_" attachment format to support multiple attachments
So this commit creates _ap2_ entries.  It currently also still creates _ap_ entries as well.
_ap_ format will still be used if no _ap2_ entries are found, _ap2_ will be ignored by earlier version of opensim
This code appears to work with viewer 3 but it not yet well tested with other viewers.
The data formats are not final - do not use in any context where you value your appearance until arrangements become final.